Chat/Mention popup
Creates floating text entity above the player with the chat message content, if this contains a player-name of a online player it creates a floating text entity above them to notify them they have been mentioned. This is useful for people that are not focusing on the chat.

BUGS:
Text not centered when zooming out, could be changed that the text is not centered but better get some player feedback to see if they even like the feature.
Text longer than 128 characters overlaps when zooming out.